additional data retrived for archived entities
