+++ server/spotify/helpers.ts
@@ -1,9 +1,7 @@
 import axios, { AxiosError } from "axios";
 import SpotifyWebApi from "spotify-web-api-node";
-import { getConnection } from "typeorm";
-import HistoryArtist from "../db/entities/HistoryArtist";
-import HistoryTrack from "../db/entities/HistoryTrack";
-import User from "../db/entities/User";
+import { archiveHistory } from "./archiveHelpers";
+
 
 const spotifyApi = new SpotifyWebApi({
   clientId: process.env.CLIENT_ID,
@@ -16,10 +14,10 @@ const getRecentlyPlayed = async (access_token: string, user_id: string) => {
   spotifyApi.setAccessToken(access_token);
   return await spotifyApi
     .getMyRecentlyPlayedTracks({
-      limit: 25,
+      limit: 2,
     })
     .then((data) => {
-      archiveHistory(data, user_id);
+      archiveHistory(data, user_id, access_token);
       // Output items
       return data;
     })
@@ -27,112 +25,6 @@ const getRecentlyPlayed = async (access_token: string, user_id: string) => {
       console.log("Error from getRecentlyPlayed", error);
     });
 };
-
-const archiveHistory = async (data: any, user_id: string) => {
-  try {
-    const { items } =  data.body;
-    return Promise.all(items.map( (trackObj: any) => {
-        return createHistoryTrack(trackObj, user_id)
-      }
-    )).then((tracks) => {
-      createHistoryArtist(tracks, user_id);
-    });
-  } catch (err) {
-    console.error('Error archiving data\'s body!');
-  };
-  return;
-}
-
-const createHistoryTrack = async (trackObj: any, user_id: string) => {
-  const { played_at, track } = trackObj;
-  const { name, uri, album } = track;
-
-  const historyTrack = await HistoryTrack.findOne({where:{user_id: user_id, played_at: played_at, track_title: name}});
-  if(historyTrack === undefined) {
-    const artists = track.album.artists.map((artist: any) => {
-      return artist.name;
-    });
-
-    const newTrack = new HistoryTrack();
-    newTrack.user_id = user_id;
-    newTrack.track_title = name;
-    newTrack.played_at = played_at;
-    newTrack.track_uri = uri;
-    newTrack.artists = artists;
-    newTrack.album_title = album.name;
-    newTrack.album_art = album.images[1].url;
-    newTrack.album_uri = album.uri;
-    await newTrack.save();
-
-    await getConnection()
-      .createQueryBuilder()
-      .relation(User, "historyTracks")
-      .of(user_id)
-      .add(newTrack);
-    return track; 
-  }
-  return;
-}
-
-const createHistoryArtist = async (tracks: Array<any>, user_id: string) => {
-  let artistObj: any = {};
-  tracks.forEach((track) => {
-    if(track){
-      if(artistObj.hasOwnProperty(track.artists[0].name)){
-        artistObj[track.artists[0].name].count += 1;
-        artistObj[track.artists[0].name].time_listened += track.duration_ms;
-      } else {
-        const tempArtist = {
-          artist_name: track.artists[0].name,
-          artist_uri: track.artists[0].uri,
-          time_listened: track.duration_ms,
-          count: 1
-        }
-        artistObj[track.artists[0].name] = tempArtist;
-      }
-    }
-  });
-  const artistArray: Array<string> = Object.values(artistObj);
-
-  try {
-    Promise.all(artistArray.map(
-      (artist) => {
-       console.log(artist)
-       createArtist(artist, user_id);
-      }
-    ))
-  } catch (err) {
-    console.error('Error archiving artists!', err);
-  };
-}
-
-const createArtist = async(artistObj: any, user_id: string) => {
-  const { artist_uri, artist_name, count, time_listened } = artistObj;
-  let historyArtist = await HistoryArtist.findOne({where:{user_id: user_id, artist_name: artist_name}});
-  if(historyArtist) {
-    await getConnection()
-      .createQueryBuilder()
-      .update(HistoryArtist)
-      .set({count: () => `count + ${count}`})
-      .set({count: () => `time_listened + ${time_listened}`})
-      .where("user_id = :user_id AND artist_name = :artist_name", {user_id: user_id, artist_name: artist_name})
-      .execute();
-  } else {
-    let newArtist = new HistoryArtist();
-    newArtist.artist_name = artist_name;
-    newArtist.artist_uri = artist_uri;
-    newArtist.user_id = user_id;
-    newArtist.time_listened = time_listened
-    newArtist.count = count;
-    await newArtist.save();
-    return await getConnection()
-      .createQueryBuilder()
-      .relation(User, "historyArtists")
-      .of(user_id)
-      .add(newArtist);
-  }
-  
-}
 
 const getUsersCurrentPlayback = async (access_token: string) => {
   const getCurrentPlayback: any = {
@@ -215,18 +107,6 @@ const querySpotify = (query: string, access_token: string) => {
   });
 };
 
-
-const getArtistData = async (artist_uri: string, access_token: string) => {
-  return await axios({
-    url: `https://api.spotify.com/v1/artists/${artist_uri}`,
-    method: "get",
-    headers: {
-      Authorization: `Bearer ${access_token}`,
-    },
-  }).then(({data}) => {
-    return data.genres;
-  }).catch(error => console.log(error));
-};
 const addToPlaylist = async (
   access_token: string,
   playlist_id: string,
@@ -292,5 +172,4 @@ export {
   querySpotify,
   addToPlaylist,
   createPlaylist,
-  getArtistData,
 };
